+++ src/main/java/io/vn/dungxnd/duckabary/core/borrow_management/BorrowException.java
@@ -1,0 +1,7 @@
+package io.vn.dungxnd.duckabary.core.borrow_management;
+
+public class BorrowException extends RuntimeException {
+    public BorrowException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}