+++ src/main/java/io/vn/dungxnd/duckabary/core/library_management/LibraryCmdService.java
@@ -75,7 +75,6 @@ public class LibraryCmdService extends LibraryService {
         return documents;
     }
 
-
     @Override
     public UserCmdService getUserService() {
         return (UserCmdService) super.getUserService();