+++ src/main/java/io/vn/dungxnd/duckabary/core/user_management/UserManagement.java
@@ -7,11 +7,11 @@ public class UserManagement {
     private final ArrayList<AdminUser> adminUsers;
     private final UserDatabaseManagement userDatabaseManagement;
 
+    private final Object lock = new Object();
+
     /**
-     * Constructor for UserManagement
-     * Initialize the list of users and admin users
-     * Load users and admin users from the database
-     * Add the loaded users and admin users to the list
+     * Constructor for UserManagement Initialize the list of users and admin users Load users and
+     * admin users from the database Add the loaded users and admin users to the list
      */
     public UserManagement() {
         this.adminUsers = new ArrayList<>();
@@ -22,12 +22,20 @@ public class UserManagement {
     }
 
     public User getUserById(int userId) {
-        return users.get(userId - 1);
+        return users.stream()
+                .filter(user -> user.id() == userId - 1)
+                .findFirst()
+                .orElseThrow(() -> new UserNotFoundException(userId));
     }
 
     public void createUser(User user) {
-        users.add(user);
-        userDatabaseManagement.addUserToDB(user);
+        synchronized (lock) {
+            if (checkUsernameExist(user.username())) {
+                throw new IllegalArgumentException("Username already exists");
+            }
+            users.add(user);
+            userDatabaseManagement.addUserToDB(user);
+        }
     }
 
     public int getNewUserID() {
@@ -39,7 +47,8 @@ public class UserManagement {
     }
 
     public void createAdminUser(int id, String username, String email, String rawPassword) {
-        AdminUser adminUser = new AdminUser(getNewAdminUserID(), username, email, rawPassword);
+        AdminUser adminUser =
+                AdminUser.createAdmin(getNewAdminUserID(), username, email, rawPassword);
         adminUsers.add(adminUser);
     }
 
@@ -58,7 +67,7 @@ public class UserManagement {
 
     public boolean checkUsernameExist(String username) {
         for (User user : users) {
-            if (user.getUsername().equals(username)) {
+            if (user.username().equals(username)) {
                 return true;
             }
         }
@@ -67,5 +76,27 @@ public class UserManagement {
 
     public UserDatabaseManagement getUserDatabaseManagement() {
         return userDatabaseManagement;
+    }
+
+    public boolean updateUser(User updatedUser) {
+        int index = findUserIndexById(updatedUser.id());
+
+        if (index != -1) {
+            users.set(index, updatedUser);
+
+            userDatabaseManagement.updateUserInDB(updatedUser);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    private int findUserIndexById(int userId) {
+        for (int i = 0; i < users.size(); i++) {
+            if (users.get(i).id() == userId) {
+                return i;
+            }
+        }
+        return -1;
     }
 }