+++ src/main/java/io/vn/dungxnd/duckabary/core/user_management/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package io.vn.dungxnd.duckabary.core.user_management;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(int userId) {
+        super("User not found with ID: " + userId);
+    }
+}