+++ src/main/java/io/vn/dungxnd/duckabary/core/user_management/UserService.java
@@ -9,16 +9,44 @@ public class UserService {
         this.userManager = userManager;
     }
 
-    public void createUser(
+    public User createUser(
             String username,
             String email,
             String firstName,
             String lastName,
             String phone,
             String address) {
-        int userId = getNewUserID();
-        User user = new User(userId, username, firstName, lastName, email, phone, address);
+        validateUserInput(username, email, phone, address);
+        int userId = userManager.getNewUserID();
+        User user = User.createUser(userId, username, firstName, lastName, email, phone, address);
         userManager.createUser(user);
+        return user;
+    }
+
+    protected void validateUserInput(String username, String email, String phone, String address) {
+        if (username == null || username.isBlank()) {
+            throw new IllegalArgumentException("Username cannot be empty");
+        }
+
+        if (isUsernameAlreadyExist(username)) {
+            throw new IllegalArgumentException("Username already exists");
+        }
+
+        if (username.length() < 5 || username.length() > 20) {
+            throw new IllegalArgumentException("Username must be between 5 and 20 characters");
+        }
+
+        if (email == null || !email.matches("^[A-Za-z0-9+_.-]+@(.+)$")) {
+            throw new IllegalArgumentException("Invalid email format");
+        }
+
+        if (phone == null || !phone.matches("^[0-9]{8,11}$")) {
+            throw new IllegalArgumentException("Invalid phone number");
+        }
+
+        if (address == null || address.isBlank()) {
+            throw new IllegalArgumentException("Address cannot be empty");
+        }
     }
 
     public int getNewUserID() {
@@ -44,4 +72,7 @@ public class UserService {
         return userManager.getUsers();
     }
 
+    public void updateUser(User user) {
+        userManager.updateUser(user);
+    }
 }