+++ src/main/java/io/vn/dungxnd/duckabary/db/DatabaseException.java
@@ -1,0 +1,7 @@
+package io.vn.dungxnd.duckabary.db;
+
+public class DatabaseException extends RuntimeException {
+    public DatabaseException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}